Update 3.1
Updates to the MCC Page.

SERVICES PAGE
The development of the service request form started and was uploaded.

PARTNERS PAGE
fixed the horizontal overflow of the page in mobile view.
-Scaled down the size of the star ratings from 1 to 0.98 and scaled up the social icons from 1 to 1.13 for the mobile view.
-Changed the height of the partner slider from 100px to 85px in the mobile view.
changed the partner banner to a slider in the mobile view.
Page responsiveness is included and is still being optimized.

CONTACT PAGE
-made all fields a requirement to avoid missing data from clients.